requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
bcrypt==4.1.1
jwt==1.3.1
PyJWT==2.8.0
werkzeug==3.0.1
sqlalchemy==2.0.23
